+++ CHANGELOG.md
@@ -8,7 +8,8 @@
   - Hides the Reuse's creation card and community resource's creation card on dataset view.
 - Newsletter's subscription banner [#512](https://github.com/etalab/udata-gouvfr/pull/512):
   - Override alert section in posts list template to display newsletter's subscription banner.
-  - Add settings `NEWSLETTER_LINK` and `NEWSLETTER_SUB_MESSAGE` to configure subscription's link and message.
+  - Add settings `POST_BANNER_ACTIVATED` to toggle banner's activation.
+  - Add settings `POST_BANNER_LINK` and `POST_BANNER_MESSAGE` to configure subscription's link and message.
 
 ## 2.2.3 (2020-11-09)